feat(git): Pass the refined git base to the Configuration (#2634)
As discussed in #2622, it does not make much sense to pass the
"incorrect"* git base around. Indeed, due to this, we need to get the
refined git base whenever we need to do a git diff, i.e. when we get the
list of changed files or the modified lines, we need to refine the base
used.

Instead, in this PR, we refine the base first when it is relevant (i.e.
we use a git diff). Thanks to this, the downstream code can be
simplified.

I also renamed "git base branch" to "git base" as it is no longer the
branch name necessarily, it can be the branch short name, full name or
hash.

*: incorrect here can designate several things: the reference not being
resolved (e.g. using `main` instead of the specific commit hash), or
having the wrong reference resolved (see #2633).