refactor(git): Make DiffChangedLinesParser I/O-less (#2642)
While inlining `DiffChangedLinesParser` as per #2622, I stumbled on the
issue that it uses `realpath()`, which meant that the docs of the method
were incorrect (and fixed in #2641).

It may be a bit controversial, but I think it's the wrong place where to
introduce this filesystem dependency. This would allow the git adapter
to remain more agnostic. The reason why we need this real path is due to
the usage in `FilesDiffChangedLines`, and it would change if
`FilesDiffChangedLines` takes a non-absolute path, so IMO it makes more
sense for `FilesDiffChangedLines` to take care of that task.