fix(git): Do not include invalid changed lines ranges (#2649)
Related to #2622.

In theory, the raw git diff lines may include changes for which there is
no changed line in the new file. If that happens, currently, we create a
`ChangedLinesRange` object nonetheless, which is incorrect.

Whilst it is a fix, it currently has no impact for the users. Indeed, we
apply the git filter `AM`. For this (the buggy) scenario to occur, you
need to have and old file for which the lines are deleted, without any
other change. This cannot happen with that filter.

Nonetheless, this is still worth fixing as the filter may (and is
likely) to change in the future.