docs(git): Document our decision on not couting a renamed file as having all of its lines changed (#2662)
I stumbled upon this case in ##2658 and wanted to shed some light on it.

As a reminder, currently if a file is renamed, none of its lines appears
in the diff meaning there is no line to mutate. There was a question of
whether this is correct.

We decided that we cannot see a case where we would want to change that
behaviour. Whilst the test is already there, this PR adds a comment to
document that this is a conscious decision.