feat(PHPUnit): Leverage `--exclude-source-from-xml-coverage` with PHPUnit 12.5+ (#2604)
As of [PHPUnit 12.5.0](https://github.com/sebastianbergmann/phpunit/releases/tag/12.5.0), it is possible to omit the `<source>` element in the XML coverage reports by passing the option `--exclude-source-from-xml-coverage`.

This drastically speeds up the generation process of the coverage report (~15% speedup on PHPStan) and should significantly reduce the size of the XML reports.

---------

Co-authored-by: Théo FIDRY <5175937+theofidry@users.noreply.github.com>