refactor(SourceCollector): Move the collection of files outside of the ConfigurationFactory (#2672)
This is the very first part of #2380 (besides some minor prep around
it).

This PR moves the collection of files out of the `ConfigurationFactory`.

<hr/>

For additional context in case you do not want to read the entirety of
#2380.

The general idea of #2380 is that we will do less processing in
`ConfigurationFactory` (which requires to know how to collect files and
filters), and rather forward the necessary information to the
`Configuration`.

In this PR, we limit ourselves to defer the collection to a service
which consumes the configuration information. Ultimately, the same
should happen for the Git filter, and then the filters should be
directly given to the `SourceFileCollector` rather than having another
API on top of it.