feat: Explicitly require `psr/log` (#2677)
Previously, we depended on `psr/log`, but this dependency was only
constrained indirectly by `composer/xdebug-handler` and
`phpbench/phpbench` (the latter being a dev dependency).

There is a break with `psr/log` 2.x regarding the tests, so as of 2.x
you also need to require `fig/log`. To avoid this annoyance, I putted
the requirement as 2.x minimum.

So TL:DR, this PR:

- Add an explicit requirement on `psr/log` which was otherwise already
required.
- Bumps the minimum to drop 1.x.