refactor: Improve the ChangedLinesRange API (#2667)
- Introduce `::forLine()` and `::forRange()` which make it a bit more
expressive and allow more documentation & self-documentation.
- Introduce `::create()` for internal usage with validation and make
`__construct()` private. The inputs are already validated for
`::forLine()` and `::forRange()` hence there is no need to have a double
validation (which is what would happen if we out the guard in the
constructor).
- Introduces a method `::touches()` to simplify the code and add more
tests for it too.
- Simplify the tests of `FilesDiffChangedLines`: they focus more on the
fact that we deal with multiple changed lines and multiple files rather
than on the definition of "the mutation touches a range of changed
lines".
- Renamed `FilesDiffChangedLines::contains()` to
`FilesDiffChangedLines::touches()`. We ask the differ if the mutation
touches any of the changed lines.