refactor(SourceCollector) Merge FileFilter with SourceCollector (#2678)
This PR inlines `SourceFileFilter` into `SourceFileCollector`.

As per #2380, the goal is to make the collectors aware of the filters to
apply so that they can collect the files in a more optimized way (e.g.
for a git diff, you can directly apply the filter) instead of enforcing
"collect all + apply the filter".