refactor(SourceCollector): Introduce a factory for BufferedSourceFilter (#2679)
The original intent of this PR was to introduce a `::create()` static
factory which makes it easier to leverage promoted properties. My hope
was to make this class readonly. Unfortunately, it is not. I kept the
factory because I find it a bit nicer like that, but I admit it is
subjective.

I made the factory accept the collected sources files rather than
passing it the collector, this is a bit more standard I think.