refactor: Merge the various filters into one as `SourceFilter` (#2395)
This PR introduces `SourceFilter`, for which the implementations are a
value object containing the relevant information. As such, we currently
have:

- `PlainFilter`, which corresponds to the value of `--filter`. I later
plan to move the parsing of its values from
`SourceFileCollector::parseFilter()` to `PlainFilter::parse()`.
- `IncompleteGitDiffFilter` which does **NOT** implement the interface.
This represents the information passed by the user (`--git-diff-filter`,
`--git-diff-lines`, `--git-diff-base`). It does not implement the
interface because the information it holds is incomplete.
- `GitDiffFilter`, the completed version of `IncompleteGitDiffFilter`.

This allows to directly expression the filter we apply in one property
`$sourceFilter`, rather than 4 different parameters with some that are
mutually exclusive (e.g. can filter with git if we already have the
regular filter) and provides a safer API for parameters that go together
(if we have a git filter, we need a base).

To keep things simple, this PR converts the introduced `GitDiffFilter`
into a `PlainFilter` for the `Configuration`. That allows us to still
work with the current `SourceFileCollector`. The plan however, via
#2380, is to make `SourceFileCollector` and have a git implementation
that will consume the `GitDiffFilter`. This will also allow us to remove
the git related properties from `Configuration`.