refactor: Make PlainFilter null if contains no values (#2684)
Transform `PlainFilter::create()` to `::tryToCreate()`. This allows us
to harden the type of `PlainFilter#values` to be a non-empty array.

Side note: "tryCreate" would be more idiomatic in C#/.net, Java or JS,
but I don't find that omitting the "to" brings that much clarity and
makes the code less expressive IMO.