feat(SourceCollector): Introduce SourceLineMatcher (#2689)
This PR introduces the `SourceLineMatcher` interface:

```php
namespace Infection\Source\Matcher\SourceLineMatcher;

/**
 * Determines whether a specific line range in a file matches certain criteria.
 */
interface SourceLineMatcher
{
    /**
     * Checks whether the specified line range in the given file matches the
     * criteria defined by this matcher.
     *
     * @param string $fileRealPath Absolute path to the file to check.
     * @param positive-int $startLine Starting line number of the range (inclusive).
     * @param positive-int $endLine Ending line number of the range (inclusive).
     *
     * @throws NoSourceFound
     */
    function touches(string $fileRealPath, int $startLine, int $endLine): bool;
}
```

This interfaces matches the signature of `FilesDiffChangedLines`. The
latter now implements that interface and has been renamed/re-allocated
accordingly.

Rationale for the naming:

- It is about checking if something matches a condition or pattern, not
filtering. Hence "matcher" is a better fit than "filter" IMO. What is
about filtering here is its _usage_, we do use it to filter out nodes to
mutate. But the method itself is not.
- I picked "source line" rather than "files diff". "diff" here is only
because the data is coming from a git diff, but that is an
implementation detail. I find "source" a bit nicer than "files",
subjective I guess.

Having an interface for this allows us to introduce a
`NullSourceLineMatcher` which can be used when there is no git filter
used. This allows us to get rid of this awful hack that was:

```php
return (new ReflectionClass(FilesDiffChangedLines::class))->newInstanceWithoutConstructor();
```

Which was necessary because we always needed a `FilesDiffChangedLines`
even if the latter was not being used.

This also allows us to get rid of the abstraction leak
`NodeMutationGenerator#isForGitDiffLines`: whether a diff is applied or
not is determined by the implementation.