fix(SourceCollector): Do not fail when no executed code is found for a filtered source (#2691)
When we execute infection with a git diff, if no line of source code is
executed, we do not fail the process. This is handled by
`IndexXmlCoverageParser` which throws a different `NoSourceFound`
exception (different exception message, and with the property
`#isSourceFiltered: bool`).

However, when the user executes infection with `--filter`, this logic
does not apply. IMO this is incorrect: the concept is exactly the same,
we work with a subset of the configured source and we do not want to
fail because it just happens that there is nothing to mutate for that
subset.

This PR fixes this and adjusts the naming of `IndexXmlCoverageParser` to
reflect this.