refactor(SourceCollector): Simplify the SourceCollectorFactory (#2701)
I did not like much how `SourceCollectorFactory` got complicated and
harder to test in #2697.

In #2699 I kinda auto-piloted in a similar solution. But actually, I
think it makes more sense to delegate that wiring of the laziness to the
container.

`SourceCollectorFactory` is about creating the right source collector
based on the filter used. The rest: laziness, caching, etc. are how we
want to wire the different services, which IMO is better off left to the
container configuration.

This also means if tomorrow the container supports laziness, the factory
won't change. If you need the factory for tests but don't want any of
the laziness or caching, this will also be possible.