feat: Introduce the Tracer (#2707)
This PR completes a 6 years old challenge:
https://github.com/infection/infection/pull/927

We introduced back then `Trace`, which is an envelope for a source file
which contains all the necessary data about its tests. The `Tracer` is
the service that provides such a trace object for a given source file.

## Motivation

Currently we work as follow:

- go through the coverage data (the XML coverage index, in
`PhpUnitXmlCoverageTraceProvider`)
- for each tests, find the source the test is about
- filter out the tests for which the source is not part of the collected
sources (this is `CoveredTraceProvider` using
`BufferedSourceFileFilter`)
- if we allow non-covered code, we create a `ProxyTrace` for each source
file that was not found in the tests (this is `UncoveredTraceProvider`
using `BufferedSourceFileFilter`)
- for each trace, start the mutation generation process

One one hand this makes sense because we start from the data we have:
the coverage. On another, it seems a bit weird to start from the
coverage which may contain a lot more than necessary, especially if we
request specific source files (via filters).

In [Architecture
PoC](https://github.com/infection/infection/pull/2272/files#diff-9bc3105e506d7103f2b25138650f3e784a373d04b529f12ff34136a2fccc72c6),
the proposal is to:

- parse the source file
- enrich the AST (mark what code is eligible, dry areas, add types,
etc.)
- starting the mutation generation process

So for that purpose, the current architecture is unfit as we start from
the wrong end. The `Tracer` allows to inverse the control flow: we
request the data (the `Trace`) for a given source file, not the other
way around.


## The solution

The API of the Tracer is fairly simple, and in all implementation
attempts, it always ended up with the same API.

The implementation idea draws inspiration from some experiments I did
when working on the coverage data in
https://github.com/infection/infection/pull/2543. The idea is:

- make the trace provider into a generator
- to lookup the information, we traverse the generator, caching the
results as we go
- once we found the desired result, we pause the traverse and yield that
result
- once we request more information, if it is not in the cache, we resume
the traversal

The current implementation may feel somewhat redundant and a bit of a
patchwork, and that is because it is. The plan is to "inline"
`TraceProvider` entirely (I started some work towards that in #2543). It
changes nothing to the general principle however.

Some benchmarks were conducted in
https://github.com/infection/infection/pull/2708#issuecomment-3659713580
and I extended the current tracing benchmarks.

The TL:DR; is that this should not result in any performance
degradation. The gains are quite unpredictable by their nature, but the
idea is that this API will _enable_ gains. See the PR for a more
detailed explanation.