test(e2e): Move the artefacts of the test for some e2e tests (#2730)
I want to introduce some improvements in #2624, hence I focused on the
tests I'm working on and not all the e2e tests.

What I propose is to move the temporary artifact to a `var` directory to
allow to keep the test a bit cleaner even when executed locally.

This is loosely related to
https://github.com/infection/infection/issues/2498.