test: Improve the tests of FileMutationGenerator (#2737)
I wanted to capture additional scenarios for the fix of #2734, but I was
really stuck at the tests. This PR rewrites them in a way which is
hopefully more understandable.

- The test is split in two: there was an integration test in there which
I extracted, this allows to have most of the tests not marked as
`integration`.
- In introduced a tiny `Scenario` object.
- I consolidated the path in two:
- One which validates the happy path and is quite strict about the
arguments passed etc.
- The other which is where we are more interested in knowing if we're
skipping the mutation generation or not and where the scenario object is
done. Since we tightly tested the flow previously, this test is written
a bit lighter and makes the assumption traverse called = mutations
generated.