fix: Checks for the trace steps before proceeding with the parsing of a file (#2734)
Fixes an issue introduced in #2707.

In there I did the mistake of conflicting the existence of a trace with
the existence of tests. We previously (before #2707) were checking the
existence of tests for the trace, this should remain.

This PR introduces this back. Thanks to #2737, it is easy to understand
the difference in tests.