feat(Tracer): Make the tracer throw a dedicated exception when no trace is found (#2740)
This enforces the exception thrown in #2739 at the contract level.

This allows us to simplify the `FileMutationGenerator` so that instead
of having to deal with both `hasTrace()` and a potential exception, we
only have to deal with the exception.

This will also allow us to further simplify the `Tracer` implementation
down (next PR).