refactor(Tracer): Remove the unused BufferedSourceFileFilter (#2745)
Previously as the `TraceProvider` was the primary source for the
mutation generations, we had to filter out the traces generated for
files that were not part of the collected files. This was done via
`BufferedSourceFileFilter`, which also allowed to expose the files from
the collected files that were not seen in the code coverage.

The flow is now different however. The primary source is the collected
source files (leveraging `SourceCollector` in `FileMutationGenerator`),
so we know that if we get a trace from the `Tracer`, it is for a source
file, there is no additional filtering to be done.

What is still true is that `CoveredTraceProvider` still goes through all
the files from the coverage report, so it may still go through files
that do not belong to the collected source files. However, in that case,
the `Tracer` will store it but not use it (it makes no difference that
we are able to actively not store or not) and the generated traces are
super lightweight (the additional processing like adding the execution
info from the JUnit report is done lazily).