feat(SourceCollector): Fail earlier when no source file is found (#2703)
## Problem

Currently, if no source file is found, either because the configured
source in `infection.json5` results in no file or because of the
`--filter` applied, then Infection still executes the initial test run.

## Solution

This PR changes the contract of `SourceCollector` to document the throw
of `NoSourceFound` and make `BasicSourceCollector` throw if no source is
found. Now, the above scenario will fail before the initial test run is
attempted. Examples of exceptions thrown:

```shell
# no filter applied
In NoSourceFound.php line 98:
No source file found for the configured sources.

# filter applied
In NoSourceFound.php line 98:
No source file found for the filter applied to the configured sources. The filter used was: "src/Unknown1,src/Unknown2".
```

The exception thrown respects whether a filter is applied, i.e. in the
2nd scenario we would would terminate with a success error code.

Note that this behaviour was already present if the user used
`--git-diff-lines` or `--git-diff-filter`.
`GitDiffSourceCollector::collect()` was failing (actually it is
`::create()`, but in practice we lazily instantiate this service hence
this only happens when calling `::collect()`).

## Caveats & change of behaviours

### Change 1)

One (very minor) issue is due to our simplistic implementation of
`GitDiffSourceCollector`, in which we re-use `BasicSourceCollector`, we
could have a slightly confusing exception message. Indeed, if the git
diff itself returns changed files, but none of them are part of the
configured source, then we would end up with the message from above. It
could be improved by mentioning information about the git info we are
using (like we do if no file is found from git diff).

This could either be fixed by re-implementing `GitDiffSourceCollector`
in a less lazy way or or catching the exception and rethrowing it with a
corrected message, but it is not the scope of this PR.

### Change 2)

Maybe obvious, but it now means that where previously the test were
executed regardless of whether a filter was applied or any source code
was configured, this is now no longer true.