feat(git): Add git commands (#2669)
This PR adds a bunch of commands essentially exposing our Git API to the
CLI.

The intent here is to somewhat ease debugging in some scenarios. You may
get away with a `var_dump();die;` statement sometimes, but it is
sometimes not possible (e.g. when using the PHAR).

Now, one there is two big concerns here:

- Maintainability: do we want to main them. It's not that much code,
neither is it critical by any means. But maybe we don't want to have
that around. Personally I would like more such utility commands for
various debugging purposes, but maybe you do not.
- One big limitation here is that I do not load the configuration, hence
the git filters displayed are not exact. I would like to change that,
but in its current form it just feels too... tedious. This does not make
those diff commands entirely useless, but does limit it a bit.