test: Rework the PHPUnit end-to-end tests (#2624)
## Backgorund

For #2543 I need to touch a lot of the coverage part and one thing I am
missing is getting hands on a good enough sample of coverage for various
test frameworks and different versions of those.

We've had a mix of fixtures:

- I introduced some for the tracing test in
`/Users/tfidry/Project/Humbug/infection/tests/phpunit/TestFramework/Tracing/Fixtures`
in ##2444.
- We have some comprehensive ones with a bit more information about what
wee get from the coverage data with
`/Users/tfidry/Project/Humbug/infection/tests/phpunit/Fixtures/TestFramework/PhpUnit/Coverage/*`
(see #957)

I'm however not too happy with the current state of things: it's hard to
find the right balance to maintain them.


## Solution

The idea of this PR is to leverage the end-to-end tests. We have e2e
tests for the various test frameworks with their different
setup/versions. What we can do is to actually enrich them a bit to be
able to generate coverage from them, when needed.

We don't really need to go out of our way to automate it: getting the
coverage for PHPUnit 9.3 for instance doesn't need to be done more. The
day we need more out of it, e.g. we want to see how a trait is described
in the XML coverage or something so we need to add a new class/tests.
Then we can pick up the generated coverage and place it where needed.

## Implementation

This PR updates the code to offer a better range of covered and
non-covered code as well as a `Makefile` command to make the test
generation easy.

I plan to touch the PhpSpec & Codeception in a similar fashion too.